+++ runtime/hetdesrun/trafoutils/io/load.py
@@ -211,7 +211,7 @@ def transformation_revision_from_python_code(code: str) -> TransformationRevisio
                         else InputType.REQUIRED
                     ),
                 )
-                for input_name, input_info in component_info_dict["inputs"].items()
+                for input_name, input_info in component_info_dict["inputs"].items()  # type: ignore
             ],
             outputs=[
                 TransformationOutput(
@@ -225,7 +225,7 @@ def transformation_revision_from_python_code(code: str) -> TransformationRevisio
                     # input info maybe a datatype string (backwards compatibility)
                     # or a dictionary containing the datatype
                 )
-                for output_name, output_info in component_info_dict["outputs"].items()
+                for output_name, output_info in component_info_dict["outputs"].items()  # type: ignore
             ],
         ),
         content=code,
@@ -246,13 +246,7 @@ def load_transformation_revisions_from_directory(  # noqa: PLR0912
         for file in files:
             path = os.path.join(root, file)
             ext = os.path.splitext(path)[1]
-            if ext not in (".py", ".json"):
-                logger.warning(
-                    "Invalid file extension '%s' to load transformation revision from: %s",
-                    ext,
-                    path,
-                )
-                continue
+
             if ext == ".py":
                 logger.info("Loading transformation from python file %s", path)
                 python_code = load_python_file(path)
@@ -268,7 +262,7 @@ def load_transformation_revisions_from_directory(  # noqa: PLR0912
                         )
                         continue
 
-            if ext == ".json":
+            elif ext == ".json":
                 logger.info("Loading transformation from json file %s", path)
                 transformation_json = load_json(path)
                 try:
@@ -276,6 +270,14 @@ def load_transformation_revisions_from_directory(  # noqa: PLR0912
                 except ValueError as err:
                     logger.error("ValueError for json from path %s:\n%s", download_path, str(err))
                     continue
+            else:
+                if ext != ".pyc":
+                    logger.warning(
+                        "Invalid file extension '%s' to load transformation revision from: %s",
+                        ext,
+                        path,
+                    )
+                continue
             transformation_dict[transformation.id] = transformation
             if ext == ".py":
                 if transform_py_to_json: