+++ runtime/tests/sql_adapter/conftest.py
@@ -106,6 +106,8 @@ def temporary_prefilled_sqlite_ts_db(
 
     del_test_df.to_sql("deletion_test_table", engine, if_exists="replace", index=False)
 
+    engine.dispose()
+
     return temporary_sqlite_file_path_ts_db
 
 
@@ -134,6 +136,9 @@ def two_sqlite_dbs_configured(temporary_sqlite_file_path, _clean_configured_dbs_
         ],
     ) as _fixture:
         yield _fixture
+
+    for c in _fixture:
+        c.engine.dispose()
 
 
 @pytest.fixture(scope="function")  # noqa: PT003
@@ -186,6 +191,9 @@ def three_sqlite_dbs_configured(
         ],
     ) as _fixture:
         yield _fixture
+
+    for c in _fixture:
+        c.engine.dispose()
 
 
 @pytest.fixture(scope="session")