+++ runtime/tests/test_backend_transformation_router.py
@@ -21,6 +21,7 @@ from hetdesrun.persistence.models.transformation import TransformationRevision
 from hetdesrun.trafoutils.filter.params import FilterParams
 from hetdesrun.trafoutils.io.load import (
     load_json,
+    load_python_file,
     transformation_revision_from_python_code,
 )
 from hetdesrun.utils import State, get_uuid_from_seed
@@ -1927,7 +1928,7 @@ async def test_execute_for_nested_workflow(async_test_client, mocked_clean_test_
             "./transformations/components/basic/filter_100_18260aab-bdd6-af5c-cac1-7bafde85188f.json",
             "./transformations/components/basic/greater-or-equal_100_f759e4c0-1468-0f2e-9740-41302b860193.json",
             "./transformations/components/basic/last-datetime-index_100_c8e3bc64-b214-6486-31db-92a8888d8991.json",
-            "./transformations/components/basic/restrict-to-time-interval_100_bf469c0a-d17c-ca6f-59ac-9838b2ff67ac.json",
+            "./transformations/components/basic/restrict-to-time-interval_100_bf469c0a-d17c-ca6f-59ac-9838b2ff67ac.py",
             "./transformations/components/connectors/pass-through-float_100_2f511674-f766-748d-2de3-ad5e62e10a1a.json",
             "./transformations/components/visualization/single-timeseries-plot_100_8fba9b51-a0f1-6c6c-a6d4-e224103b819c.json",
             "./transformations/workflows/examples/data-from-last-positive-step_100_2cbb87e7-ea99-4404-abe1-be550f22763f.json",
@@ -1935,18 +1936,17 @@ async def test_execute_for_nested_workflow(async_test_client, mocked_clean_test_
             "./transformations/workflows/examples/linear-rul-from-last-positive-step_100_3d504361-e351-4d52-8734-391aa47e8f24.json",
         ]
 
-        for file in json_files:
-            tr_json = load_json(file)
+        tr_list = [
+            load_json(file) if file.endswith(".json") else load_python_file(file)
+            for file in json_files
+        ]
 
-            response = await ac.put(
-                posix_urljoin(
-                    get_config().hd_backend_api_url,
-                    "transformations",
-                    tr_json["id"],
-                ),
-                params={"allow_overwrite_released": True},
-                json=tr_json,
-            )
+        response = await ac.put(
+            posix_urljoin(get_config().hd_backend_api_url, "transformations"),
+            params={"allow_overwrite_released": True},
+            json=tr_list,
+        )
+        assert response.status_code == 207
 
         component_id = UUID("57eea09f-d28e-89af-4e81-2027697a3f0f")
         updated_component = read_single_transformation_revision(component_id)
@@ -2094,19 +2094,26 @@ async def test_put_workflow_transformation(async_test_client, mocked_clean_test_
         "./transformations/components/basic/filter_100_18260aab-bdd6-af5c-cac1-7bafde85188f.json",
         "./transformations/components/basic/greater-or-equal_100_f759e4c0-1468-0f2e-9740-41302b860193.json",
         "./transformations/components/basic/last-datetime-index_100_c8e3bc64-b214-6486-31db-92a8888d8991.json",
-        "./transformations/components/basic/restrict-to-time-interval_100_bf469c0a-d17c-ca6f-59ac-9838b2ff67ac.json",
+        "./transformations/components/basic/restrict-to-time-interval_100_bf469c0a-d17c-ca6f-59ac-9838b2ff67ac.py",
         "./transformations/components/connectors/pass-through-float_100_2f511674-f766-748d-2de3-ad5e62e10a1a.json",
     ]
 
-    for file in json_files:
-        tr_json = load_json(file)
-        store_single_transformation_revision(TransformationRevision(**tr_json))
-
-    example_workflow_tr_json = load_json(
-        "./transformations/workflows/examples/data-from-last-positive-step_100_2cbb87e7-ea99-4404-abe1-be550f22763f.json"
-    )
+    tr_list = [
+        load_json(file) if file.endswith(".json") else load_python_file(file) for file in json_files
+    ]
 
     async with async_test_client as ac:
+        response = await ac.put(
+            posix_urljoin(get_config().hd_backend_api_url, "transformations"),
+            params={"allow_overwrite_released": True},
+            json=tr_list,
+        )
+        assert response.status_code == 207
+
+        example_workflow_tr_json = load_json(
+            "./transformations/workflows/examples/data-from-last-positive-step_100_2cbb87e7-ea99-4404-abe1-be550f22763f.json"
+        )
+
         response = await ac.put(
             posix_urljoin("/api/transformations/", example_workflow_tr_json["id"]),
             json=example_workflow_tr_json,