+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * remove deprecated web endpoints for base-items, components, workflows, wirings and documentation
 * remove exporting capabilities for exporting from the pre 0.7 Java backend
+* allow to change single operator revision to a DRAFT revision
 
 ## 0.12.1
 * dependency upgrades and upgrade to Python 3.13